Scripted Board Data
Developed a method that writes the board data into a text file (for use
later with an editor app). Need to build the reverse of this process, to
generate boards from text files... in the end, these text files will be
generated by the editor app with a gui for tile selection.